% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CalculateEffectiveNiche_new}
\alias{CalculateEffectiveNiche_new}
\title{CalculateEffectiveNiche}
\usage{
CalculateEffectiveNiche_new(object, cutoff = 0.05)
}
\arguments{
\item{object}{A niche-DE object}

\item{cutoff}{Minimum kernel similarity. Similarities below this value get truncated to 0}
}
\value{
A niche-DE object the effective niche calculated.
The effective niche is a list with each entry corresponding to a kernel bandwidth
}
\description{
This function calculates the effective niche of a niche-DE object
}
