% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CreateLibraryMatrix}
\alias{CreateLibraryMatrix}
\title{CreateLibraryMatrix}
\usage{
CreateLibraryMatrix(data, cell_type)
}
\arguments{
\item{data}{Single cell rna-seq counts matrix. Dimension should be cells/spots by genes}

\item{cell_type}{Cell_type assignment matrix. First column is cell names and second column is cell type assignment.}
}
\value{
A library matrix with the average expression profile per cell type
}
\description{
This function creates a expression profile matrix for single cell data
}
