% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CreateLibraryMatrixFromSeurat}
\alias{CreateLibraryMatrixFromSeurat}
\title{CreateLibraryMatrixFromSeurat}
\usage{
CreateLibraryMatrixFromSeurat(seurat_object, assay)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{assay}{The assay from which we want to extract the counts matrix to
calculate the average expression profile.}
}
\value{
A library matrix with the average expression profile per cell type
}
\description{
This function creates a expression profile matrix for single cell data
from a Seurat object
}
