% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CreateNicheDEObject}
\alias{CreateNicheDEObject}
\title{CreateNicheDEObject}
\usage{
CreateNicheDEObject(
  counts_mat,
  coordinate_mat,
  library_mat,
  deconv_mat,
  sigma,
  Int = T
)
}
\arguments{
\item{counts_mat}{Counts matrix. Dimension should be cells/spots by genes}

\item{coordinate_mat}{Coordinate matrix. It will be scaled such that the median nearest neighbor distance is 100.}

\item{library_mat}{Matrix indicating average expression profile for each cell type in the sample}

\item{deconv_mat}{Deconvolution or cell type assignment matrix of data}

\item{sigma}{List of kernel bandwidths to use in calculating the effective niche}

\item{Int}{Boolean of if counts data supplied is integer. Default is true. When performing niche-DE,
Negative binomial regression is performed if True. Linear regression with a gene specific variance is performed if False.}
}
\value{
A niche-DE object
}
\description{
This function creates a niche-DE object
}
