% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CreateNicheDEObjectFromSeurat}
\alias{CreateNicheDEObjectFromSeurat}
\title{CreateNicheDEObjectFromSeurat}
\usage{
CreateNicheDEObjectFromSeurat(
  seurat_object,
  assay,
  library_mat,
  deconv_mat,
  sigma
)
}
\arguments{
\item{seurat_object}{A spatial seurat object.Coordinate matrix will be extracted via the
seurat function 'GetTissueCoordinates'. The coordiantes will be scaled such that the median nearest neighbor distance is 100.}

\item{assay}{The assay from which to extract the counts matrix from. The counts matrix
will be extracted from the counts slot.}

\item{library_mat}{Matrix indicating average expression profile for each cell type in the sample}

\item{deconv_mat}{Deconvolution or cell type assignment matrix of data}

\item{sigma}{List of kernel bandwidths to use in calculating the effective niche}
}
\value{
A niche-DE object
}
\description{
This function creates a niche-DE object from a seurat object
}
