% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{Filter_NDE}
\alias{Filter_NDE}
\title{Filter_NDE}
\usage{
Filter_NDE(object, cell_names)
}
\arguments{
\item{object}{A niche-DE object}

\item{cell_names}{Cell names of observations that should be kept}
}
\value{
A niche-DE object that only includes the specified cells.
We recommend using this function after calculating the effective niche on the whole dataset
}
\description{
This function filters a niche-DE object to only include specific observations
}
