% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_main_functions.R
\name{get_niche_DE_genes}
\alias{get_niche_DE_genes}
\title{Getniche genes for the given index and niche cell types at the desired test.level}
\usage{
get_niche_DE_genes(
  object,
  test.level,
  index,
  niche,
  positive = T,
  alpha = 0.05
)
}
\arguments{
\item{object}{A niche-DE object}

\item{test.level}{At which test.level to return genes
(gene level ("G"), cell type level("CT"), interaction level("I"))}

\item{index}{The index cell type}

\item{niche}{The niche cell type}

\item{positive}{Logica indicating whether to return upregulated/positive niche-DE genes(True) or downregulated/negative niche-DE genes(False). Default value is True.}

\item{alpha}{The level at which to perform the Benjamini Hochberg correction. Default value = 0.05}
}
\value{
A vector of genes that are niche significant at the desired FDR,
test.level, index cell type, and niche cell type
}
\description{
This function returns genes that show niche patterns at the desired test.level
}
