% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_main_functions.R
\name{get_niche_DE_pval}
\alias{get_niche_DE_pval}
\title{Get Niche-DE Pvalues}
\usage{
get_niche_DE_pval(object, pos = T)
}
\arguments{
\item{object}{A niche-DE object}

\item{pos}{Logical indicating whether to calculate pvalues for(index,niche)+
patterns (pos = T) or (index,niche)- patterns (pos = F)}
}
\value{
A niche-DE object with niche-DE pvalues
}
\description{
This function calculates pvalues from a niche-DE analysis
}
