% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{MergeObjects}
\alias{MergeObjects}
\title{MergeObjects}
\usage{
MergeObjects(objects)
}
\arguments{
\item{objects}{A list of niche-DE objects}
}
\value{
A niche-DE object with each niche-DE object that concatenates
each niche-DE object used in the input list. Coordinates are scaled
to ensure that kernel bandwidths are consistent across datasets
}
\description{
This function merges niche-DE objects
}
