% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_main_functions.R
\name{niche_DE_markers_new}
\alias{niche_DE_markers_new}
\title{Get Niche-DE marker genes(test)}
\usage{
niche_DE_markers_new(object, index, niche1, niche2, alpha = 0.05)
}
\arguments{
\item{object}{A niche-DE object}

\item{index}{The index cell type which we want to find marker genes for}

\item{niche1}{The niche cell type for the marker genes found}

\item{niche2}{The niche we wish to compare (index,niche1) patterns to}

\item{alpha}{The level at which to perform the Benjamini Hochberg correction. Default value is 0.05.}
}
\value{
A vector of genes that are niche marker genes for the index cell type
near the niche1 cell type relative to the niche2 cell type
}
\description{
This function returns genes marker genes in the index cell type when near the first niche cell type realtive to the second one
}
