% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_object_creation.R
\name{CalculateEffectiveNicheLargeScale}
\alias{CalculateEffectiveNicheLargeScale}
\title{CalculateEffectiveNiche}
\usage{
CalculateEffectiveNicheLargeScale(
  object,
  batch_size = 1000,
  cutoff = 0.05,
  standardize = TRUE
)
}
\arguments{
\item{object}{A niche-DE object}

\item{batch_size}{Number of cells to calculate the effective niche for at a time}

\item{cutoff}{Minimum kernel similarity. Similarities below this value get truncated to 0}

\item{standardize}{A logical indicator for if the effective niche should be z-normalized by column}
}
\value{
A niche-DE object the effective niche calculated.
The effective niche is a list with each entry corresponding to a kernel bandwidth
}
\description{
This function calculates the effective niche of a niche-DE object
}
