% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_DE_main_functions.R
\name{get_niche_DE_genes}
\alias{get_niche_DE_genes}
\title{Get niche genes for the given index and niche cell types at the desired resolution}
\usage{
get_niche_DE_genes(object, resolution, index, niche, pos, alpha)
}
\arguments{
\item{object}{A niche-DE object}

\item{resolution}{At which resolution to return genes
(gene level, cell type level, interaction level)}

\item{index}{The index cell type}

\item{niche}{The niche cell type}

\item{pos}{Logical indicating whether to return genes that are (index,niche)+
patterns (pos = T) or (index,niche)- (pos = F)}

\item{alpha}{The level at which to perform the Benjamini Hochberg correction}
}
\value{
A vector of genes that are niche significant at the desired FDR,
resolution, index cell type, and niche cell type
}
\description{
This function returns genes that show niche patterns at the desired resolution
}
